Fix: Migrate server routes and controllers
