+++ backend-server/api/routes/app.route.ts
@@ -1,8 +1,16 @@
 import express, { IRouter } from 'express';
-import { getAppController } from '../controllers/app.controller';
+import { createIssueController, getAppController, getIssuesController, getPullRequestsController, getRepositoriesController } from '../controllers/app.controller';
 
 const router: IRouter = express.Router();
 
 router.get('/', getAppController);
+//-------------------------------------------
+router.get('/issues', getIssuesController);
+//-------------------------------------------
+router.post('/issues', createIssueController);
+//-------------------------------------------
+router.get('/pull-requests', getPullRequestsController);
+//-------------------------------------------
+router.get('/repositories', getRepositoriesController);
 
 export { router };
